Fix bug detail page: Remove incorrect field swapping that caused 404 errors
The bug detail page had a 'TEMPORARY FIX' that was swapping bugId and title fields.
This caused the page to use the bug title in the URL instead of the bug ID, resulting
in 404 errors when trying to view bug details.

Fixed by:
- Removed the field swapping logic in loadBugData()
- Bug data is now used directly as returned from the API
- Bug ID is correctly used in all API calls and URLs

This resolves the issue where clicking 'View Details' on a bug would result in:
GET /api/bugs/Test%20Bug%20-%20Email%20Notification 404 (Not Found)